linux-requirements.txt
requests
django
djangorestframework
markdown
django-debug-toolbar
cherrypy
whitenoise
croniter
django-q2
rich
python-dateutil
jinja2
# defusedxml
# django-guardian
django_bootstrap5
pydantic
xmltodict
python-json-logger
django_extensions
records
mariadb